+++ packages/components/ui/rollup.config.js
@@ -48,7 +48,7 @@ const input = {
 	IconX: 'src/icons/IconX.svelte',
 	index: 'src/index.js',
 	LinkButton: 'src/LinkButton.svelte',
-	ScreenGauge: 'src/gauges/ScreenGauge.svelte',
+	ScreenGauge: 'src/gauges/screen/ScreenGauge.svelte',
 	Switch: 'src/Switch.svelte',
 };
 const removeComments = cleanup({